+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Adding tests for CLI argument parsing for the `-e` flag.
+
 ## [0.3.11] - 2025-10-25
 
 ### Added